requirements.txt
django==5.0.6
django-taggit==5.0.1
numpy==1.26.4
scipy==1.14.1
scikit-learn==1.5.2
transformers==4.44.2
torch==2.4.1
django-cloudinary-storage==0.3.0 
cloudinary==1.36.0 
whitenoise==6.6.0 
gunicorn==21.2.0 
psycopg2-binary==2.9.9 
pillow==10.1.0 
sentencepiece==0.1.99